\name{hello}
\alias{hello}
\title{Useful functions in data analysis}
\usage{
hello()
}
\description{
Prints 'Hello, world!'.
}
\examples{
hello()
}



\usage{
charFactor()
}
\description{
 Change all char vars into factors
}
\examples{
charFactor(data)
}
\usage{
getCor()
}
\description{
 get Corrmatrix of all numberic cols
}
\examples{
getCor(data)
}




\usage{
missValue(data)
}
\description{
 get the num of NAs in each col
}
\examples{
missVale(data)
}

